pyproject.toml
[tool.poetry]
name = "edu-service"
version = "0.1.0"
description = ""
authors = ["gennady <topchiev@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
pillow = "^10.4.0"
djangorestframework = "^3.15.2"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
ipython = "^8.26.0"
drf-spectacular = "^0.27.2"
drf-yasg = "^1.21.7"
setuptools = "^70.3.0"
django-cors-headers = "^4.4.0"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.0"
black = "^24.4.2"
isort = "^5.13.2"

[tool.black]
line-length = 120
extend-exclude = """
    (
    .*migrations/
    | .*tests/
    )
"""

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
